Consolidate prefixing node labels with deleted to its own function. Move to deletes file.
