+++ src/core/database/query/deletes.ts
@@ -1,6 +1,7 @@
 import { stripIndent } from 'common-tags';
 import { node, Query, relation } from 'cypher-query-builder';
 import { DateTime } from 'luxon';
+import { ResourceShape } from '../../../common';
 
 export const deleteBaseNode = (query: Query) =>
   query
@@ -30,24 +31,59 @@ export const deleteBaseNode = (query: Query) =>
        we need to distinct propertyNode to avoid collecting and labeling each propertyNode more than once
        */
     .with('[baseNode] + collect(propertyNode) as nodeList')
-    /**
-       check if labels already have the "Deleted_" prefix to avoid "Deleted_Deleted_"
-       yielding a node from the label procedures is necessary I believe
-       they're not used in rest of the query and are aliased to avoid colliding with the unwound "node" alias
-       */
-    .raw(
-      stripIndent`
-        unwind nodeList as node
-        with node,
-        reduce(
-          deletedLabels = [], label in labels(node) |
-            case
-              when label starts with "Deleted_" then deletedLabels + label
-              else deletedLabels + ("Deleted_" + label)
-            end
-        ) as deletedLabels
-        call apoc.create.removeLabels(node, labels(node)) yield node as nodeRemoved
-        with node, deletedLabels
-        call apoc.create.addLabels(node, deletedLabels) yield node as nodeAdded
-      `
-    );
+    .raw('unwind nodeList as node')
+    .apply(prefixNodeLabelsWithDeleted('node'));
+
+/**
+ * This will set all relationships given to active false
+ * and add deleted prefix to its labels.
+ */
+export const deleteProperties =
+  <Resource extends ResourceShape<any>>(
+    _resource: Resource,
+    ...relationLabels: ReadonlyArray<keyof Resource['prototype']>
+  ) =>
+  (query: Query) => {
+    if (relationLabels.length === 0) {
+      return query;
+    }
+    const deletedAt = DateTime.local();
+    return query
+      .match([
+        node('node'),
+        relation('out', 'propertyRel', relationLabels, { active: true }),
+        node('property', 'Property'),
+      ])
+      .setValues({
+        'property.deletedAt': deletedAt,
+        'propertyRel.active': false,
+      })
+      .apply(prefixNodeLabelsWithDeleted('property'));
+  };
+
+export const prefixNodeLabelsWithDeleted = (node: string) => (query: Query) =>
+  query.subQuery((sub) =>
+    sub
+      .with([
+        node,
+        // Mpa current labels to have deleted prefix (operation is idempotent).
+        stripIndent`
+          reduce(
+            deletedLabels = [], label in labels(${node}) |
+              case
+                when label starts with "Deleted_" then deletedLabels + label
+                else deletedLabels + ("Deleted_" + label)
+              end
+          ) as deletedLabels
+        `,
+      ])
+      // yielding node is necessary even though unused
+      .raw(
+        `call apoc.create.removeLabels(${node}, labels(${node})) yield node as nodeRemoved`
+      )
+      .with([node, 'deletedLabels']) // Is this really needed?
+      .raw(
+        `call apoc.create.addLabels(${node}, deletedLabels) yield node as nodeAdded`
+      )
+      .return('1')
+  );
